Token entity renamed for better clarity
