+++ src/index.ts
@@ -1,6 +1,6 @@
 export { default as HatchConfig } from './models/HatchConfig'
 export { default as Contribution } from './models/Contribution'
-export { default as Token } from './models/Token'
+export { default as ERC20Token } from './models/ERC20Token'
 export { default as Hatch } from './models/Hatch'
 export { default as HatchConnectorTheGraph } from './thegraph/connector'
 export { default } from './connect'
