+++ src/models/HatchOracleConfig.ts
@@ -1,14 +1,14 @@
 import { HatchOracleConfigData } from '../types'
-import Token from './Token'
+import ERC20Token from './ERC20Token'
 
 class HatchOracleConfig {
   readonly id: string
-  readonly scoreToken: Token
+  readonly scoreToken: ERC20Token
   readonly ratio: number
 
   constructor(data: HatchOracleConfigData) {
     this.id = data.id
-    this.scoreToken = data.scoreToken
+    this.scoreToken = new ERC20Token(data.scoreToken)
     this.ratio = data.ratio
   }
 }
